product_apis

+++ miniStore.Application/Catalog/Products/ManageProductService.cs
@@ -7,12 +7,16 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
-using Microsoft.EntityFrameworkCore;
 using System.Linq;
 using Microsoft.AspNetCore.Http;
 using System.IO;
 using System.Net.Http.Headers;
 using miniStore.Application.Common;
+using Microsoft.EntityFrameworkCore;
+using miniStore.ViewModels.Catalog.ProductImages;
+using Microsoft.AspNetCore.Server.IISIntegration;
+using Microsoft.Data.SqlClient;
+
 namespace miniStore.Application.Catalog.Products
 {
     public class ManageProductService : IManageProductService
@@ -24,6 +28,30 @@ namespace miniStore.Application.Catalog.Products
             _context = context;
             _storageService = storageService;
         }
+        public async Task<ProductViewModel> GetById(int productId, string languageId)
+        {
+            var product = await _context.Products.FindAsync(productId);
+            var productTranslation = await _context.ProductTranslations.FirstOrDefaultAsync(x => x.ProductId == productId
+            && x.LanguageId == languageId);
+
+            var productViewModel = new ProductViewModel()
+            {
+                Id = product.Id,
+                DateCreated = product.DateCreated,
+                Description = productTranslation != null ? productTranslation.Description : null,
+                LanguageId = productTranslation.LanguageId,
+                Details = productTranslation != null ? productTranslation.Details : null,
+                Name = productTranslation != null ? productTranslation.Name : null,
+                OriginalPrice = product.OriginalPrice,
+                Price = product.Price,
+                SeoAlias = productTranslation != null ? productTranslation.SeoAlias : null,
+                SeoDescription = productTranslation != null ? productTranslation.SeoDescription : null,
+                SeoTitle = productTranslation != null ? productTranslation.SeoTitle : null,
+                Stock = product.Stock,
+                ViewCount = product.ViewCount
+            };
+            return productViewModel;
+        }
         public async Task<int> Create(ProductCreateRequest request)
         {
             var product = new Product()
@@ -33,7 +61,7 @@ namespace miniStore.Application.Catalog.Products
                 Stock = request.Stock,
                 ViewCount = 0,
                 DateCreated = DateTime.Now,
-                ProductTranslations=new List<ProductTranslation>()
+                ProductTranslations = new List<ProductTranslation>()
                 {
                     new ProductTranslation()
                     {
@@ -48,8 +76,9 @@ namespace miniStore.Application.Catalog.Products
                 },
             };
             //save image
-            if (request.ThumbnailImage != null) {
-                product.ProductImages = new List<ProductImage>() { 
+            if (request.ThumbnailImage != null)
+            {
+                product.ProductImages = new List<ProductImage>() {
                     new ProductImage(){
                         Caption="Thumbnail image",
                         DateCreated = DateTime.Now,
@@ -61,17 +90,23 @@ namespace miniStore.Application.Catalog.Products
                 };
             }
             _context.Products.Add(product);
-            return await _context.SaveChangesAsync();
+            await _context.SaveChangesAsync();
+            return product.Id;
         }
 
         public async Task<int> Delete(int productId)
         {
-            var product= await _context.Products.FindAsync(productId);
+            var product = await _context.Products.FindAsync(productId);
             if (product == null) throw new MiniStoreException($"Cannot find a product {productId}");
 
-            var images = product.ProductImages.Where(x => x.ProductId == productId);
-            foreach (var image in images) { 
-                await _storageService.DeleteFileAsync(image.ImagePath);
+            var images = product.ProductImages?.Where(x => x.ProductId == productId).ToList();
+
+            if (images != null && images.Any())
+            {
+                foreach (var image in images)
+                {
+                    await _storageService.DeleteFileAsync(image.ImagePath);
+                }
             }
 
             _context.Products.Remove(product);
@@ -81,7 +116,7 @@ namespace miniStore.Application.Catalog.Products
         public async Task<int> Update(ProductUpdateRequest request)
         {
             var product = await _context.Products.FindAsync(request.Id);
-            var productTranslations = await _context.ProductTranslations.FirstOrDefaultAsync(x => x.ProductId == request.Id && 
+            var productTranslations = await _context.ProductTranslations.FirstOrDefaultAsync(x => x.ProductId == request.Id &&
             x.LanguageId == request.LanguageId);
             if (product == null || productTranslations == null) throw new MiniStoreException($"Cannot find a product with Id: {request.Id}");
 
@@ -94,7 +129,8 @@ namespace miniStore.Application.Catalog.Products
 
             //save image
             var thumbnailImage = await _context.ProductImages.FirstOrDefaultAsync(x => x.IsDefault == true && x.ProductId == request.Id);
-            if (thumbnailImage != null) {
+            if (thumbnailImage != null)
+            {
                 thumbnailImage.FileSize = request.ThumbnailImage.Length;
                 thumbnailImage.ImagePath = await this.SaveFile(request.ThumbnailImage);
                 _context.ProductImages.Update(thumbnailImage);
@@ -119,16 +155,18 @@ namespace miniStore.Application.Catalog.Products
                         join c in _context.Categories on pic.CategoryId equals c.Id
                         select new { p, pt, pic };
 
-            if (!string.IsNullOrEmpty(request.Keyword)) { 
-                query =query.Where(x => x.pt.Name.Contains(request.Keyword));
+            if (!string.IsNullOrEmpty(request.Keyword))
+            {
+                query = query.Where(x => x.pt.Name.Contains(request.Keyword));
             }
-            if (request.CategoryId.Count() > 0) {
+            if (request.CategoryId.Count() > 0)
+            {
                 query = query.Where(x => request.CategoryId.Contains(x.pic.CategoryId));
             }
 
-            int totalRow= await query.CountAsync();
+            int totalRow = await query.CountAsync();
 
-            var data= await query.Skip((request.PageIndex - 1)* request.PageSize)
+            var data = await query.Skip((request.PageIndex - 1) * request.PageSize)
                 .Take(request.PageSize)
                 .Select(
                     x => new ProductViewModel()
@@ -159,7 +197,7 @@ namespace miniStore.Application.Catalog.Products
         public async Task<bool> UpdatePrice(int productId, decimal newPrice)
         {
             var product = await _context.Products.FindAsync(productId);
-            if (product == null)  throw new MiniStoreException($"Cannot find a product with id {productId}");
+            if (product == null) throw new MiniStoreException($"Cannot find a product with id {productId}");
             product.Price = newPrice;
             return _context.SaveChanges() > 0;
         }
@@ -179,5 +217,85 @@ namespace miniStore.Application.Catalog.Products
             return fileName;
         }
 
+        public async Task<int> AddImage(int productId, ProductImageCreateRequest request)
+        {
+            var productImage = new ProductImage()
+            {
+                Caption = request.Caption,
+                DateCreated = DateTime.Now,
+                IsDefault = request.IsDefault,
+                ProductId = productId,
+                SortOrder = request.SortOrder,
+            };
+            if (request.ImageFile != null)
+            {
+                productImage.ImagePath = await this.SaveFile(request.ImageFile);
+                productImage.FileSize = request.ImageFile.Length;
+            }
+            _context.ProductImages.Add(productImage);
+            await _context.SaveChangesAsync();
+            return productImage.Id;
+        }
+
+        public async Task<int> UpdateImage(int imageId, ProductImageUpdateRequest request)
+        {
+            var productImage = await _context.ProductImages.FindAsync(imageId);
+            if (productImage == null)
+                throw new MiniStoreException($"Cannot find an image with id:{imageId}");
+
+            if (request.ImageFile != null)
+            {
+                productImage.ImagePath = await this.SaveFile(request.ImageFile);
+                productImage.FileSize = request.ImageFile.Length;
+            }
+             _context.ProductImages.Update(productImage);
+            return await _context.SaveChangesAsync();
+        }
+
+        public async Task<int> DeleteImage(int imageId)
+        {
+            var productImage = await _context.ProductImages.FindAsync(imageId);
+            if (productImage == null)
+                throw new MiniStoreException($"Cannot find an image with id:{imageId}");
+
+            _context.ProductImages.Remove(productImage);
+            return await _context.SaveChangesAsync();
+
+        }
+
+        public async Task<ProductImageViewModel> GetImageById(int imageId)
+        {
+            var image = await _context.ProductImages.FindAsync(imageId);
+            if (image == null)
+                throw new MiniStoreException($"Cannot find an image with id:{imageId}");
+
+            var viewModel = new ProductImageViewModel()
+            {
+                Id = image.Id,
+                ProductId=image.ProductId,
+                Caption=image.Caption,
+                DateCreated=image.DateCreated,
+                FileSize=image.FileSize,
+                ImagePath=image.ImagePath,
+                SortOrder=image.SortOrder,
+                IsDefault=image.IsDefault,
+            };
+            return viewModel;
+        }
+
+        public async Task<List<ProductImageViewModel>> GetListImage(int productId)
+        {
+            return await _context.ProductImages.Where(i=>i.ProductId == productId)
+                .Select(x => new ProductImageViewModel(){
+                    Id = x.Id,
+                    ProductId=x.ProductId,  
+                    Caption=x.Caption,
+                    DateCreated=x.DateCreated,
+                    FileSize=x.FileSize,
+                    ImagePath=x.ImagePath,
+                    SortOrder=x.SortOrder,
+                    IsDefault=x.IsDefault,
+                }).ToListAsync();
+        }
     }
 }
